+++ examples/webmail/tests.py
@@ -1,6 +1,5 @@
 '''Test twisted integration'''
 import pulsar
-from pulsar import is_failure, is_async
 from pulsar.utils.pep import to_bytes, to_string
 from pulsar.apps.test import unittest, dont_run_with_thread